+++ CHANGELOG.md
@@ -37,6 +37,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### bug fixes
+
+- [#780](https://github.com/babylonlabs-io/babylon/pull/780) crypto: remove Verify for ECDSA
+
 ## v1.0.0-rc10
 
 ### Improvements