+++ CHANGELOG.md
@@ -37,6 +37,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+## v1.0.0-rc11
+
 ### bug fixes
 
 - [#780](https://github.com/babylonlabs-io/babylon/pull/780) crypto: remove Verify for ECDSA